+++ objects.cpp
@@ -19,7 +19,7 @@ GitObject* GitObject::make_object( const Pack *pack, off_t header_index, const s
 
   uint8_t byte = pack->byte( index++ );
   uint8_t type = (byte >> 4) & 7;
-  size_t size = byte & 0x0f;
+  off_t size = byte & 0x0f;
   uint8_t shift = 4;
 
   while ( byte & 0x80 ) {
@@ -62,7 +62,7 @@ void Ofs_Delta::decode_delta_ptr( void )
 {
   off_t i = data_index;
   uint8_t byte = pack->byte( i++ );
-  size_t negative_offset = byte & 0x7f;
+  off_t negative_offset = byte & 0x7f;
   while ( byte & 0x80 ) {
     byte = pack->byte( i++ );
     negative_offset = ((negative_offset+1) << 7) + (byte & 0x7f);
